Move drop predicate, drop all and schema mutations out of transaction.
